refactor: update leaderboard routes to support beatmapsets with difficulty selection
- Replaced `/leaderboard/beatmaps/[beatmapId]` with `/leaderboard/beatmapsets/[beatmapsetId]/[beatmapId]`.
- Added support for difficulty selection and preview playback for beatmapsets in leaderboards.
- Updated layout and components for compatibility with the new hierarchy.
- Improved error handling and loading state management for new routes.
